Move classes to separate source files.
- classes GenericAssembly and AssemblyBase
- prepare for using in other equations
